Cache pre-compiled `bits/stdc++.h` headers
A significant part of compiling simple C++ programs is taken up by the
parsing and analysis of the enormous `bits/stdc++.h` header. GCC's
pre-compiled headers (PCH) feature allows most of this work to be cached
by dumping this data out to a `.gch` file. These files are tied to both
the exact compiler flags as well as the specific compiler binary.
Therefore, a dynamic caching method is used which generates the PCH if
one with the appropriate key isn't available.

A later commit will add support for limiting the maximum cache size.

This will be used in the IDE's executor service, but this could be added
to the njudge website too.