Add basic size quota to the PCH cache
This ensures that the PCH cache's size does not exceed a pre-set size
limit. For simplicity, we don't track accesses like in a LRU cache, but
instead evict the oldest PCH file if we hit the quota. This is fine, as
in the case of Algo Pro IDE's execute functionality, the compiler flags
are hard-coded, so the only scenario where multiple PCHs are generated
is if the compilers are updated.